ai/pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "LLM query tool that processes file contents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
]
